# _build-openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Chat Management Api
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: Local Server run on port 8080
paths:
...
  /api/conversations:
    post:
      summary: Create Conversation
      operationId: create conversation
      description: create new conversation
      tags:
        - conv
      security:
        - bearerAuth: []
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: 'APP:SUCCESS | ERR:VALIDATION'
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/CreateConvRes'
        '400':
          $ref: '#/components/responses/Other400'
        default:
          $ref: '#/components/responses/Other200'
...
components:
...
  schemas:
...
    CreateConvRes:
      type: object
      properties:
        id:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: '2024-08-21T21:31:41.342336882+07:00'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Other200:
      description: other than 200
      content:
        application/json:
          schema:
            type: object
            required:
              - code
              - status
              - message
              - data
            properties:
              code:
                type: string
                example: 'ERR:BAD_REQUEST | ERR:VALIDATION | ERR:AUTHENTICATION'
              status:
                type: string
                example: can be a custom message
              message:
                type: string
                example: can be a custom message
              data:
                type: object
...
    Other400:
      description: other than 400
      content:
        application/json:
          schema:
            type: object
            required:
              - code
              - status
              - message
            properties:
              code:
                type: string
                example: 'ERR:BAD_REQUEST | ERR:VALIDATION | ERR:AUTHENTICATION'
              status:
                type: string
                example: can be a custom message
              message:
                type: string
                example: can be a custom message